VoiceBox/requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.6.0
psycopg[binary,pool]>=3.1.0
databricks-sdk>=0.20.0
